+++ components/SpotInfo/index.js
@@ -11,27 +11,15 @@ import {
   Longitude,
   Latitude,
   EntryList,
-  EntryCard,
-  EntryContentWrapper,
-  EntryTextarea,
   EditIcon,
-  TrashIcon,
-  EdTrButtonWrapper,
-  EntryEditErrorText,
-  ButtonAndErrorWrapper,
-  EntryEditButton,
-  EntryDeleteButton,
   NoEntryMessage,
-  AddEntryForm,
-  AddEntryLabel,
-  AddEntryTextarea,
-  AddEntryButtonWrapper,
-  AddEntryButton,
   SpotDeleteButton,
 } from "./style";
 import LoadingSpinner from "../LoadingSpinner";
 import Error from "../Error";
-import { CiEdit, CiCircleCheck, CiTrash } from "react-icons/ci";
+import { CiEdit, CiCircleCheck } from "react-icons/ci";
+import InfoForm from "../InfoForm";
+import AddNewInfoForm from "../AddNewInfoForm"
 
 export default function SpotInfo({ spotId }) {
   const {
@@ -43,12 +31,6 @@ export default function SpotInfo({ spotId }) {
   const [isEditingSpotName, setIsEditingSpotName] = useState(false);
   const [spotNameError, setSpotNameError] = useState("");
   const [newSpotName, setNewSpotName] = useState("");
-  const [newInfo, setNewInfo] = useState("");
-  const [editMode, setEditMode] = useState(null);
-  const [originalInfo, setOriginalInfo] = useState("");
-  const [editInfo, setEditInfo] = useState("");
-  const [editError, setEditError] = useState("");
-  const [entryError, setEntryError] = useState("");
   const router = useRouter();
 
   const handleSpotNameChange = (event) => {
@@ -83,97 +65,6 @@ export default function SpotInfo({ spotId }) {
     }
   };
 
-  const handleNewEntryChange = (event) => {
-    setNewInfo(event.target.value);
-  };
-
-  const handleAddNewEntry = async (event) => {
-    event.preventDefault();
-
-    if (!newInfo.trim()) {
-      setEntryError("YOU FORGOT TO ENTER THE INFORMATION");
-      return;
-    }
-
-    const response = await fetch(`/api/spots/${spotId}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ info: newInfo }),
-    });
-
-    if (response.ok) {
-      setNewInfo("");
-      mutate();
-      setEntryError("");
-    }
-  };
-
-  const handleEditChange = (event) => {
-    setEditInfo(event.target.value);
-  };
-
-  const handleEditEntry = async (infoId) => {
-    if (editMode !== null) {
-      if (!editInfo.trim()) {
-        setEditError("YOU FORGOT TO ENTER THE INFORMATION");
-        return;
-      }
-      if (editInfo === originalInfo) {
-        setEditMode(null);
-        setOriginalInfo("");
-        return;
-      }
-
-      const response = await fetch(
-        `/api/spots/${spotId}/informations/${infoId}`,
-        {
-          method: "PUT",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ info: editInfo }),
-        }
-      );
-
-      if (response.ok) {
-        setEditInfo("");
-        setEditMode(null);
-        setOriginalInfo("");
-        mutate();
-        setEditError("");
-      } else {
-        const errorData = await response.json();
-        setEditError(errorData.message);
-      }
-    } else {
-      setEditMode(infoId);
-      const infoToBeEdited = spot.informations.find(
-        (info) => info._id === infoId
-      );
-      setEditInfo(infoToBeEdited.info);
-      setOriginalInfo(infoToBeEdited.info);
-    }
-  };
-
-  const handleDeleteEntry = async (infoId) => {
-    const response = await fetch(
-      `/api/spots/${spotId}/informations/${infoId}`,
-      {
-        method: "DELETE",
-      }
-    );
-
-    if (response.ok) {
-      const data = await response.json();
-
-      if (data && data.success) {
-        mutate();
-      }
-    }
-  };
-
   const handleDeleteSpot = async () => {
     const response = await fetch(`/api/spots/${spotId}`, {
       method: "DELETE",
@@ -185,7 +76,7 @@ export default function SpotInfo({ spotId }) {
   };
 
   if (isValidating) {
-    return <LoadingSpinner role="status"/>;
+    return <LoadingSpinner role="status" />;
   }
 
   if (error) {
@@ -211,12 +102,15 @@ export default function SpotInfo({ spotId }) {
         ) : (
           spot.spotName
         )}
-        <SpotNameEditButton onClick={handleEditSpotName} aria-label="Edit spot name">
+        <SpotNameEditButton
+          onClick={handleEditSpotName}
+          aria-label="Edit spot name"
+        >
           {isEditingSpotName ? (
             <EditIcon as={CiCircleCheck} size={25} />
           ) : (
             <EditIcon as={CiEdit} size={25} />
-          )} 
+          )}
         </SpotNameEditButton>
       </SpotName>
       <SpotNameError>{spotNameError}</SpotNameError>
@@ -230,67 +124,14 @@ export default function SpotInfo({ spotId }) {
         {spot.informations && spot.informations.length > 0 ? (
           <EntryList>
             {spot.informations.map((entry) => (
-              <EntryCard key={entry._id}>
-                <EntryContentWrapper>
-                  <EntryTextarea
-                    isEditing={editMode === entry._id}
-                    value={editMode === entry._id ? editInfo : undefined}
-                    onChange={
-                      editMode === entry._id ? handleEditChange : undefined
-                    }
-                  >
-                    {entry.info}
-                  </EntryTextarea>
-                  <EdTrButtonWrapper>
-                    <EntryEditButton onClick={() => handleEditEntry(entry._id)} aria-label={`Edit entry for ${entry.info}`}>
-                      {editMode === entry._id ? (
-                        <EditIcon as={CiCircleCheck} size={25} />
-                      ) : (
-                        <EditIcon as={CiEdit} size={25} />
-                      )}
-                    </EntryEditButton>
-                    <EntryDeleteButton
-                      onClick={() => handleDeleteEntry(entry._id)} aria-label={`Delete entry for ${entry.info}`}
-                    >
-                      <TrashIcon as={CiTrash} size={25} />
-                    </EntryDeleteButton>
-                  </EdTrButtonWrapper>
-                </EntryContentWrapper>
-                {editError && editMode === entry._id && (
-                  <ButtonAndErrorWrapper>
-                    <EntryEditErrorText aria-live="polite">{editError}</EntryEditErrorText>
-                  </ButtonAndErrorWrapper>
-                )}
-              </EntryCard>
+              <InfoForm key={entry._id} entry={entry} spotId={spotId} />
             ))}
           </EntryList>
         ) : (
           <NoEntryMessage>There is no entry yet</NoEntryMessage>
         )}
+        <AddNewInfoForm spotId={spotId}/>
       </InformationWrapper>
-      <AddEntryForm onSubmit={handleAddNewEntry} aria-label="Form to add a new entry">
-        <AddEntryLabel htmlFor="new-info">
-          ADD SOME SPOT INFORMATION
-        </AddEntryLabel>
-        <EntryEditErrorText aria-live="polite">{entryError}</EntryEditErrorText>
-        <AddEntryTextarea
-          id="new-info"
-          name="new-info"
-          maxLength="450"
-          value={newInfo}
-          onChange={handleNewEntryChange}
-          aria-label="Add new entry input"
-        />
-        <AddEntryButtonWrapper>
-          <AddEntryButton
-            type="submit"
-            name="create-info"
-            aria-label="Create entry"
-          >
-            add this entry
-          </AddEntryButton>
-        </AddEntryButtonWrapper>
-      </AddEntryForm>
       <SpotDeleteButton onClick={handleDeleteSpot}>
         Delete this Spot
       </SpotDeleteButton>
