+++ components/SpotInfo/style.js
@@ -71,7 +71,7 @@ export const Latitude = styled.p`
   margin: 0px 0px 10px 0px;
 `;
 
-export const NoEntryMessage = styled.div`
+export const NoEntryMessage = styled.p`
   text-transform: uppercase;
   font-size: 1rem;
   color: #ff0000;
@@ -84,151 +84,8 @@ export const EntryList = styled.ul`
   align-items: left;
 `;
 
-export const EntryCard = styled.li`
-  display: flex;
-  flex-direction: column-reverse;
-  align-items: flex-start;
-  justify-content: space-between;
-`;
-
-export const EntryContentWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-`;
-
-export const ButtonAndErrorWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: 100%;
-`;
-
-export const EntryTextarea = styled(({ isEditing, ...props }) =>
-  isEditing ? <textarea {...props} /> : <p {...props} />
-)`
-  flex: 1;
-  list-style: none;
-  font-family: Roboto;
-  padding: 10px;
-  height: auto;
-  width: 100%;
-  background-color: #ffffff;
-  border-radius: 5px;
-  margin-bottom: 10px;
-  max-width: 90%;
-  max-height: 100%;
-  overflow-wrap: break-word;
-  word-wrap: break-word;
-  white-space: pre-wrap;
-
-  ${({ isEditing }) =>
-    isEditing
-      ? `
-    border: dotted 1.5px #5d9ea6;
-    
-      `
-      : `
-    background-color: #fcfcfc;
-    border: solid 1px #d5d5d5;
-      `}
-  &:focus {
-    background-color: #ffffff;
-    outline: solid 1px #5d9ea6;
-    border: none;
-  }
-`;
-
 export const EditIcon = styled.svg`
   fill: #5d9ea6;
-`;
-
-export const TrashIcon = styled.svg`
-  fill: #ff3232;
-`;
-
-export const EdTrButtonWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const EntryEditErrorText = styled.p`
-  color: #ff3232;
-  font-size: 0.9rem;
-  margin: 0px 0px 2px 0px;
-`;
-
-export const EntryEditButton = styled.button`
-  text-transform: uppercase;
-  background-color: transparent;
-  border: none;
-  margin: 0px 0px 5px 10px;
-  height: 40px;
-  width: 40px;
-`;
-
-export const EntryDeleteButton = styled.button`
-  text-transform: uppercase;
-  background-color: transparent;
-  border: none;
-  margin: 0px 0px 5px 10px;
-  height: 40px;
-  width: 40px;
-`;
-
-export const AddEntryForm = styled.form`
-  width: 90%;
-  display: flex;
-  flex-direction: column;
-  align-items: left;
-`;
-
-export const AddEntryLabel = styled.label`
-  font-size: 1.1rem;
-  font-weight: 500;
-  text-transform: uppercase;
-  margin: 40px 0px 20px 0px;
-`;
-
-export const AddEntryTextarea = styled.textarea`
-  font-family: Roboto;
-  background-color: #fcfcfc;
-  border: solid 1px #d5d5d5;
-  border-radius: 5px;
-  padding: 10px;
-  height: 100%;
-  width: 100%;
-
-  &:hover {
-    background-color: #f7f7f7;
-  }
-
-  &:focus {
-    background-color: #ffffff;
-    outline: solid 1px #5d9ea6;
-    border: none;
-  }
-`;
-
-export const AddEntryButtonWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 100%;
-`;
-
-export const AddEntryButton = styled.button`
-  text-transform: uppercase;
-  background-color: #fcfcfc;
-  border: solid 1px #d5d5d5;
-  border-radius: 5px;
-  margin: 15px 0px 15px 0px;
-  height: 23px;
-  width: 125px;
-
-  &:hover {
-    background-color: #f7f7f7;
-  }
 `;
 
 export const SpotDeleteButton = styled.button`
